requirements.txt
amqp==2.6.1
asgiref==3.5.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==4.4.2
certifi==2021.10.8
charset-normalizer==2.0.11
Django==4.0.2
django-phonenumber-field==6.0.0
idna==3.3
kombu==4.6.11
phonenumbers==8.12.43
pytz==2021.3
redis==3.5.3
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.8
vine==1.3.0
whitenoise==5.3.0
